data-plane-python/k8s/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.6.2
    # via httpcore
attrs==22.1.0
    # via
    #   -r requirements.in
    #   nile-api
cachetools==5.2.0
    # via google-auth
certifi==2022.9.24
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
charset-normalizer==2.1.1
    # via requests
emoji==2.1.0
    # via -r requirements.in
google-auth==2.13.0
    # via kubernetes
h11==0.12.0
    # via httpcore
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via nile-api
idna==3.4
    # via
    #   anyio
    #   requests
    #   rfc3986
kubernetes==24.2.0
    # via -r requirements.in
nile-api==0.2.3
    # via -r requirements.in
oauthlib==3.2.2
    # via requests-oauthlib
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
python-dateutil==2.8.2
    # via
    #   kubernetes
    #   nile-api
python-dotenv==0.21.0
    # via -r requirements.in
pyyaml==6.0
    # via
    #   -r requirements.in
    #   kubernetes
requests==2.28.1
    # via
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via kubernetes
rfc3986[idna2008]==1.5.0
    # via httpx
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   kubernetes
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
urllib3==1.26.12
    # via
    #   kubernetes
    #   requests
websocket-client==1.4.1
    # via kubernetes

# The following packages are considered to be unsafe in a requirements file:
# setuptools
